refactor(acp-bridge): lift status + paths + errors (#4175 PR 22b checkpoint 1)
First three slices of the PR 22b bulk lift. All purely mechanical
moves with re-export wrappers at the old locations so every existing
import keeps resolving.

What moves
- `packages/cli/src/serve/status.ts` (600 LOC) → `packages/acp-bridge/src/status.ts`
  (acpAgent.ts:85-109 imports 25 symbols from this; lifting unifies
  the wire-contract owner.)
- `packages/cli/src/serve/status.test.ts` → `packages/acp-bridge/src/status.test.ts`
- `canonicalizeWorkspace` from `cli/src/serve/fs/paths.ts:61-97` →
  `packages/acp-bridge/src/workspacePaths.ts` (cross-module contract;
  bridge package now owns it directly)
- `MAX_WORKSPACE_PATH_LENGTH` from `httpAcpBridge.ts:762` → same
  `workspacePaths.ts` file
- 11 bridge error classes from `httpAcpBridge.ts` (`SessionNotFoundError`,
  `RestoreInProgressError`, `InvalidSessionScopeError`,
  `SessionLimitExceededError`, `WorkspaceMismatchError`,
  `InvalidClientIdError`, `InvalidPermissionOptionError`,
  `InvalidSessionMetadataError`, `WorkspaceInitConflictError`,
  `McpServerNotFoundError`, `McpServerRestartFailedError`) → new
  `packages/acp-bridge/src/bridgeErrors.ts`

Wrappers
- `cli/src/serve/status.ts` shrinks to a 1-line `export *` re-export
- `cli/src/serve/fs/paths.ts` keeps its other exports
  (`resolveWithinWorkspace`, `Intent`, `ResolvedPath`,
  `hasSuspiciousPathPattern`); imports + re-exports the lifted two
- `httpAcpBridge.ts` consolidated `import` + `export` block for all 11
  error classes plus `MAX_WORKSPACE_PATH_LENGTH`; the local factory
  code keeps using these symbols since they're now imported into
  module scope

Package wiring
- `packages/acp-bridge/package.json` — adds `@qwen-code/qwen-code-core`
  dependency (status types reference `SkillError` from core); adds
  subpath exports for `./status`, `./workspacePaths`, `./bridgeErrors`,
  plus future `/bridgeTypes`, `/bridgeOptions`, `/bridge`, `/spawnChannel`
- `packages/acp-bridge/tsconfig.json` — adds `paths` mapping for
  `@qwen-code/qwen-code-core` and a project reference so typecheck
  resolves core sources directly without needing dist/ first
- `packages/cli/tsconfig.json` already has `@qwen-code/acp-bridge/*`
  path mapping from PR 22a, no change needed
- `package-lock.json` — surgical patch adding `@qwen-code/qwen-code-core`
  dep to acp-bridge's workspace entry; restored from origin/main first
  to avoid the npm-11 peer-flag churn that broke PR 22a's first push

Backward compatibility
- Every existing relative import from `cli/src/serve/` keeps working
  through wrappers (commands/serve.ts, server.ts, runQwenServe.ts,
  workspaceAgents.ts, workspaceMemory.ts, acpAgent.ts)
- Zero `/capabilities`, route, SDK, or behavior changes

Verification
- `npm ci --no-audit --ignore-scripts` clean (1453 packages)
- `cd packages/core && npx tsc --build` clean
- `cd packages/acp-bridge && npx tsc --noEmit` clean (with moved tests)
- `cd packages/cli && npx tsc --noEmit` clean

Remaining PR 22b checkpoints
- 22b/2: bridge types + BridgeOptions + DaemonStatusProvider seam
- 22b/3: BridgeClient class
- 22b/4: defaultSpawnChannelFactory
- 22b/5: createHttpAcpBridge factory (~2240 LOC of moves)
- 22b/6: serve/daemonStatusProvider.ts + runQwenServe wire
- 22b/7: move bridge tests
- 22b/8: shrink httpAcpBridge.ts to final re-export shim
