+++ packages/cli/src/serve/fs/paths.ts
@@ -4,97 +4,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { promises as fsp, realpathSync } from 'node:fs';
+import { promises as fsp } from 'node:fs';
 import * as path from 'node:path';
 import { isWithinRoot } from '@qwen-code/qwen-code-core';
 import { FsError, type FsErrorKind } from './errors.js';
 
-/**
- * Canonicalize a workspace path so the boot-time bound path and every
- * request's `workspaceCwd` collapse to the same key. `path.resolve`
- * alone normalizes `..` and `.` segments and absolutizes, but on
- * case-insensitive filesystems (macOS APFS, Windows NTFS) `/Work/A`
- * and `/work/a` are the same directory yet `resolve` returns them
- * verbatim — without normalization the `boundWorkspace` check would
- * reject every request that spelled the path with different casing
- * and `sessionScope: 'single'` re-attach would silently degrade to
- * "one per spelling".
- *
- * `realpathSync.native` (when the path exists) walks symlinks and returns
- * the on-disk casing; this matches what `config.ts` / `settings.ts` /
- * `sandbox.ts` use for their own workspace resolution. When the path
- * doesn't exist (test fixtures, ahead-of-mkdir flows) we fall back to
- * the resolved-but-uncanonicalized form rather than throwing — the
- * downstream `spawn({cwd})` will fail with a useful ENOENT if the
- * workspace truly doesn't exist.
- *
- * NOTE: This is a **cross-module contract** (BX9_q) — `config.ts`,
- * `settings.ts`, `sandbox.ts`, and `httpAcpBridge.ts` all need to
- * canonicalize the same way for the bound-workspace check +
- * `sessionScope: 'single'` re-attach to work correctly across paths.
- * The contract: use `realpathSync.native` on the resolved absolute
- * path; fall back to `path.resolve` only when the path doesn't exist
- * yet. If a future change breaks this alignment (e.g. one module
- * starts lowercasing on Windows but this one doesn't), the
- * canonicalized request path won't match the canonicalized bound
- * path → every request returns `workspace_mismatch` even though the
- * human-readable paths look equivalent. There's no test that pins
- * the alignment; the integration suite would catch a divergence only
- * if it tested the specific casing / symlink path the affected
- * module changed.
- *
- * Stage 2 in-process (#3803 §10) collapses the bridge into core,
- * removing the bridge-side path resolution entirely. Stage 1.5
- * `@qwen-code/acp-bridge` lift (chiga0 finding 1) is the natural
- * place to extract a shared `canonicalizeWorkspace` primitive that
- * all four modules consume — the lowest-common-denominator
- * extraction is fine THERE because the package boundary forces the
- * call sites to converge. Until then, *any* change to how those
- * modules resolve workspace paths needs a matching change here.
- *
- * #4175 PR 18 extraction: this file is the new home of the primitive
- * for the serve layer. The bridge re-exports it so existing callers
- * continue to import from `httpAcpBridge.js`. The forthcoming
- * `WorkspaceFileSystem` boundary (PR 18 commits 3+) builds on top of
- * this single resolver.
- */
-export function canonicalizeWorkspace(p: string): string {
-  const resolved = path.resolve(p);
-  try {
-    // FIXME(stage-2): switch to `fs.promises.realpath` once the
-    // bridge call sites become async-friendly. This sync syscall
-    // runs on the hot `spawnOrAttach` path and blocks the event
-    // loop for one filesystem stat per call. Single-user loopback
-    // (Stage 1's design target) doesn't notice; high-concurrency
-    // deployments will. Stage 2 in-process refactor removes the
-    // entire bridge-side path resolution anyway, but if Stage 2
-    // ever lands without that change, switch to the async version.
-    //
-    // Note for the resolveWithinWorkspace fast-path:
-    // `CANONICAL_BOUND_CACHE` (defined later in this file) memoizes
-    // the canonical mapping per `boundWorkspace` string. Under the
-    // `1 daemon = 1 workspace` model the cache is hit 100% of the
-    // time after the factory's boot-time canonicalization, so the
-    // sync syscall in steady state is **zero per request** —
-    // event-loop blocking only happens at boot or whenever a fresh
-    // `boundWorkspace` value first appears (e.g. tests sharing a
-    // module instance across `mkdtemp` workspaces).
-    return realpathSync.native(resolved);
-  } catch (err) {
-    // Only fall back to path.resolve for ENOENT (path doesn't exist
-    // yet). Other filesystem errors (EACCES, EIO, ELOOP) should
-    // propagate — swallowing them would hide transient I/O failures
-    // behind misleading workspace_mismatch rejections.
-    if (
-      err &&
-      typeof err === 'object' &&
-      (err as { code?: unknown }).code === 'ENOENT'
-    ) {
-      return resolved;
-    }
-    throw err;
-  }
-}
+// `canonicalizeWorkspace` and `MAX_WORKSPACE_PATH_LENGTH` lifted to
+// `@qwen-code/acp-bridge` in #4175 PR 22b — the bridge package owns the
+// cross-module workspace-canonicalization contract directly. Imported
+// here for the local `canonicalizeBoundWorkspaceCached` fast-path AND
+// re-exported so callers like `config.ts` / `settings.ts` /
+// `sandbox.ts` and the file-system service keep importing from
+// `./serve/fs/paths.js` without churn.
+import {
+  canonicalizeWorkspace,
+  MAX_WORKSPACE_PATH_LENGTH,
+} from '@qwen-code/acp-bridge/workspacePaths';
+export { canonicalizeWorkspace, MAX_WORKSPACE_PATH_LENGTH };
 
 /**
  * Branded absolute path that has passed the workspace boundary check.
